rename sabi-s3 -> plus-s3
