+++ libs/aws/plus-s3/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "sabi-s3"
+name = "plus-s3"
 version = "0.0.1"
 edition = "2018"
