+++ libs/aws/plus-s3/src/actions/error.rs
