+++ libs/aws/plus-s3/src/actions/get_object/response.rs
