+++ libs/aws/plus-s3/src/client/error.rs
