+++ libs/aws/plus-s3/src/core/error.rs
