+++ libs/aws/plus-s3/src/core/request/request_resource.rs
