+++ libs/aws/plus-s3/src/core/response/headers/e_tag.rs
