+++ libs/aws/plus-s3/src/core/response/headers/error.rs
