+++ libs/aws/plus-s3/src/core/verbs/has_bucket_scope.rs
