+++ libs/aws/plus-s3/src/core/verbs/to_endpoint.rs
