+++ libs/aws/plus-s3/src/internal/error.rs
