+++ libs/aws/plus-s3/src/internal/impl_async/error_response.rs
